refactor(app): extract EngineController from AppState
Move the transcription-engine concern out of AppState into its own
@MainActor controller: the three engine instances (WhisperKit /
Parakeet / Qwen3), the active-engine selection (activeTranscriptionEngine),
the settings -> engine language/vocabulary sync (up-front + self-rearming
reactive observer), and the launch-time model preload.

EngineController is fully self-contained — it needs only settings, so it
runs its up-front sync + arms its observer in its own init (no post-init
activate wiring). It is NOT @Observable: its engine references are lets
and consumers observe the engine instances themselves, so it exposes no
observable state of its own (same as LiveTranscriptionCoordinator). It
still uses withObservationTracking internally to watch the settings keys.

AppState exposes it as engines; the pipeline / live-transcription
coordinators read engines.activeTranscriptionEngine, the watch-start path
calls engines.syncLanguageSettings(), and the Settings UI + RPC snapshot
read engines.whisperKit etc. The menu-bar launch task calls
engines.preloadActiveModel() (moved out of the SwiftUI scene body — better
altitude and keeps the body's type-check budget in hand).

This is the final concern extraction in the AppState god-class split.
AppState shrank to ~396 lines (from ~997) and its init lost the unused
engine-injection parameters (no caller passed them). The remaining
AppState is a thin composer: it wires the seven controllers +
updateChecker + liveCaptions and exposes the derived UI properties the
menu bar binds to.
